requirements.txt
google-adk>=1.0.0
google-cloud-bigquery
google-cloud-firestore
pydantic>=2.0,<3.0
fastapi>=0.112,<1.0
uvicorn>=0.34.0,<1.0.0
firebase-admin>=6.5,<7.0
python-dotenv>=1.0,<2.0
